chore: Rename testing utilities
