+++ src/internal/analytics-metadata/index.ts
@@ -7,7 +7,6 @@ export {
   copyAnalyticsMetadataAttribute,
   getAnalyticsLabelAttribute,
 } from './attributes';
-export { getRawAnalyticsMetadata } from './test-utils';
 
 import { METADATA_DATA_ATTRIBUTE } from './attributes';
 import { GeneratedAnalyticsMetadata, GeneratedAnalyticsMetadataFragment } from './interfaces';
