+++ src/internal/analytics-metadata/testing-utils.ts
