+++ src/internal/index.ts
@@ -34,5 +34,4 @@ export {
   copyAnalyticsMetadataAttribute,
   getAnalyticsLabelAttribute,
   getGeneratedAnalyticsMetadata,
-  getRawAnalyticsMetadata,
 } from './analytics-metadata';
